Move the Configurator and StepInterface to the CoreBundle, add a compiler pass for the Configurator, add priority support to steps, deprecate existing Configurator and StepInterface implementations
