+++ app/bundles/CoreBundle/DependencyInjection/Compiler/ConfiguratorPass.php
@@ -1,0 +1,37 @@
+<?php
+/**
+ * @package     Mautic
+ * @copyright   2014 Mautic Contributors. All rights reserved.
+ * @author      Mautic
+ * @link        http://mautic.org
+ * @license     GNU/GPLv3 http://www.gnu.org/licenses/gpl-3.0.html
+ */
+
+namespace Mautic\CoreBundle\DependencyInjection\Compiler;
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\Reference;
+
+/**
+ * Class ConfiguratorPass
+ */
+class ConfiguratorPass implements CompilerPassInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function process(ContainerBuilder $container)
+    {
+        if (!$container->hasDefinition('mautic.configurator')) {
+            return;
+        }
+
+        $configuratorDef = $container->findDefinition('mautic.configurator');
+
+        foreach ($container->findTaggedServiceIds('mautic.configurator.step') as $id => $tags) {
+            $priority = isset($tags[0]['priority']) ? $tags[0]['priority'] : 0;
+            $configuratorDef->addMethodCall('addStep', array(new Reference($id), $priority));
+        }
+    }
+}
